refactor[#571] : getRandomImg 파일 위치 변경
